% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{abc}
\alias{abc}
\title{Classify a group by proportion of a variable (A,B,C)}
\usage{
abc(
  .data,
  ...,
  dim,
  a = 0.7,
  b = 0.26,
  c = 0.04,
  na.rm = TRUE,
  order_by = c("sum")
)
}
\arguments{
\item{.data}{data frame or tibble}

\item{...}{groups columns  to group by}

\item{dim}{dimension to classify by (should be additive and positive)}

\item{a}{initial segment threshold expressed as a percentage}

\item{b}{incremental segment threshold expressed as a percentage}

\item{c}{final segment threshold expressed as a percentage}

\item{na.rm}{to remove na values}

\item{order_by}{aggregate function to order by such as "sum","n","sd","mean","mad","aad","prop05","prop25","prop_50","prop_75","prop95"}
}
\value{
a tibble or obj
}
\description{
Classify a group by proportion of a variable (A,B,C)
}
\examples{

customer_abc <- abc(contoso,customer_key,dim=margin,a=.7,b=.2,c=.1)
}
