% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{abc_fn}
\alias{abc_fn}
\title{Classify a group by proportion of a variable (A,B,C,...)}
\usage{
abc_fn(x)
}
\arguments{
\item{x}{segment object}
}
\value{
a dbi objection
}
\description{
This returns a table that will segment your data into A,B or C segments based on custom
thresholds
}
