% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_graph.R
\name{abc_graph}
\alias{abc_graph}
\title{Produces ABC graph based on abc() function}
\usage{
abc_graph(df, group_label, dim_label, size, shape, ...)
}
\arguments{
\item{df}{data frame}

\item{group_label}{label for grouping variable}

\item{dim_label}{label for dimension variable}

\item{size}{ggplot size}

\item{shape}{ggplot shape}

\item{...}{}
}
\value{

}
\description{
Produces ABC graph based on abc() function
}
