% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{augment_non_standard_calendar}
\alias{augment_non_standard_calendar}
\title{Augment non-standard calendar}
\usage{
augment_non_standard_calendar(.data, pattern)
}
\arguments{
\item{.data}{data}

\item{pattern}{554,445 or 545}
}
\value{
DBI object
}
\description{
Augment non-standard calendar
}
