% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{augment_time_attributes}
\alias{augment_time_attributes}
\title{Augment time attributes}
\usage{
augment_time_attributes(.data, date_var)
}
\arguments{
\item{date_var}{}
}
\description{
Augment time attributes
}
