% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sql.R
\name{capture_original_query}
\alias{capture_original_query}
\title{Capture previous query steps}
\usage{
capture_original_query(.data)
}
\arguments{
\item{.data}{dbi object}
}
\value{
sql object
}
\description{
Capture previous query steps
}
\examples{
sales_db |> capture_original_query()
}
