% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_file_names.R
\name{clean_file_names}
\alias{clean_file_names}
\title{Rename file names witih janitor convention}
\usage{
clean_file_names(file_path, ...)
}
\arguments{
\item{file_path}{file path with files you want to rename}

\item{...}{additional args for janitor::make_clean_names()}
}
\value{
invisible
}
\description{
Rename file names witih janitor convention
}
\examples{
clean_file_names()
}
