% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{closest_sunday_feb1}
\alias{closest_sunday_feb1}
\title{Finds closet sunday to February 1st}
\usage{
closest_sunday_feb1(year)
}
\arguments{
\item{year}{year of the date}
}
\value{
character vector
}
\description{
Finds closet sunday to February 1st
}
\keyword{internal}
