% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{cohort}
\alias{cohort}
\title{Title}
\usage{
cohort(
  .data,
  .date,
  .value,
  calendar_type,
  time_unit = "month",
  period_label = FALSE
)
}
\arguments{
\item{.data}{tibble or dbi object}

\item{.date}{date column}

\item{.value}{id column}

\item{calendar_type}{standard calendar}

\item{time_unit}{do you want daily, montly, quarterly or annual view}
}
\value{
segment object
}
\description{
A database is remake of 'https://github.com/PeerChristensen/cohorts' cohort package combining
chort_table_month, cohort_table_year, cohort_table_day into a single package.
Re-written to be database friendly tested against snowflake and duckdb databases
}
