% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{cohort_fn}
\alias{cohort_fn}
\title{Internal function for the cohort segment object}
\usage{
cohort_fn(x)
}
\arguments{
\item{x}{segment object}

\item{period_label}{optional label}
}
\value{
function
}
\description{
Internal function for the cohort segment object
}
