% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{complete_calendar}
\alias{complete_calendar}
\title{complete_calendar}
\usage{
complete_calendar(x)
}
\arguments{
\item{x}{dataframe}
}
\value{
dbi object

dbi object
}
\description{
complete_calendar

complete_calendar
}
