% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{complete_non_standard_calendar}
\alias{complete_non_standard_calendar}
\title{Augment non-standard calendar}
\usage{
complete_non_standard_calendar(.data, x)
}
\arguments{
\item{.data}{data}

\item{x}{is ti object}

\item{pattern}{554,445 or 545}
}
\value{
DBI object
}
\description{
Augment non-standard calendar
}
\keyword{internal}
