% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{convert_dots_to_string}
\alias{convert_dots_to_string}
\title{Capture dot arguments and turn into strings}
\usage{
convert_dots_to_string(...)
}
\arguments{
\item{...}{dot args}
}
\value{
strings
}
\description{
Capture dot arguments and turn into strings
}
\examples{
convert_dots_to_string(hello,how,are,you)
}
