% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{convert_input_to_string}
\alias{convert_input_to_string}
\title{Convert quoted or unquoted input to string}
\usage{
convert_input_to_string(x)
}
\arguments{
\item{x}{quoted or unquoted input}
}
\value{
string
}
\description{
Convert quoted or unquoted input to string
}
\examples{
convert_input_to_string(hello)
}
