% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_plus.R
\name{count_plus}
\alias{count_plus}
\title{Augments count funiton to give proportion and cumlative proportion and sorting options (dplyr,sparklyr friendly)}
\usage{
count_plus(.data, ..., wt, .sort = 1)
}
\arguments{
\item{.data}{dataframe}

\item{...}{columns you want to groupby}

\item{wt}{if you want to weight the count (eg. value or price)}

\item{.sort}{if you want to sort from largest (1), reverse(-1) or nothing 0}
}
\value{
a tibble of values
}
\description{
Augments count funiton to give proportion and cumlative proportion and sorting options (dplyr,sparklyr friendly)
}
\examples{
ggplot2::diamonds \%>\% count_plus(color,cut,wt=price)
}
