% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sql.R
\name{create_contonso_duckdb}
\alias{create_contonso_duckdb}
\title{Create duckdb versions of Contoso datasets}
\usage{
create_contonso_duckdb()
}
\value{
DBI objects
}
\description{
Create duckdb versions of Contoso datasets
}
\examples{
create_contonso_duckdb()
}
