% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{create_non_standard_month}
\alias{create_non_standard_month}
\title{Create Non-Standard Month}
\usage{
create_non_standard_month(.data, pattern)
}
\arguments{
\item{.data}{non-standard calendar table}

\item{pattern}{'554','545' or '445'}
}
\value{
DBI object
}
\description{
Create Non-Standard Month
}
\keyword{internal}
