% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sql.R
\name{cte}
\alias{cte}
\title{CTE}
\usage{
cte(con, ...)
}
\arguments{
\item{con}{the DBI connection}

\item{...}{}
}
\value{
a list of DBI objects and SQL queries
}
\description{
CTE
}
\examples{
cte(with(sql1,sales),with(sql2,summary),dplyr::sql("select * from summary"))
}
