% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.R
\name{divide}
\alias{divide}
\title{Divide function with error handling for divide by zero or NA}
\usage{
divide(
  .data,
  new_col_name,
  numerator_col,
  denominator_col,
  alternative_result = NA_integer_
)
}
\arguments{
\item{.data}{a tibble or DBI object}

\item{new_col_name}{new column name for result}

\item{alternative_result}{alternative results if divide results are inf or NA, must be numeric}

\item{numerator}{column for numerator}

\item{denominator}{column for denominator}
}
\value{
tibble or dbi object
}
\description{
A safe divide function that will catch info or NA values and return an alternative result
This is tibble or DBI friendly and will return same class as input
}
\examples{
data(mtcars)
mtcars |> divide(new_col_name=div_col,numerator=mpg,denominator=0,alternative_result=10)
}
