% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{dod_dbi}
\alias{dod_dbi}
\title{calculate day over day values on a complete calendar table}
\usage{
dod_dbi(.data, ..., date_var, value_var, lag_n = 1)
}
\arguments{
\item{lag_n}{}
}
\description{
calculate day over day values on a complete calendar table
}
