% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{generate_cli_action}
\alias{generate_cli_action}
\title{Generate CLI actions}
\usage{
generate_cli_action(x, word)
}
\arguments{
\item{x}{input to test against}

\item{word}{the key word to validate}
}
\value{
list
}
\description{
Generate CLI actions
}
