% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{is_yyyy_mm_dd}
\alias{is_yyyy_mm_dd}
\title{Validate an input is YYYY-MM-DD format}
\usage{
is_yyyy_mm_dd(x)
}
\arguments{
\item{x}{date column}
}
\value{
logical
}
\description{
Validate an input is YYYY-MM-DD format
}
