% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{make_action_cli}
\alias{make_action_cli}
\title{Make Action field CLI args}
\usage{
make_action_cli(x)
}
\arguments{
\item{x}{}
}
\value{
list
}
\description{
Make Action field CLI args
}
\examples{
x <- ytd(fpaR::sales,.date=order_date,.value=quantity,calendar_type="standard")
make_action_cli(x@action@value)
}
