% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{make_aggregation_tbl}
\alias{make_aggregation_tbl}
\title{Aggregate and expand date table}
\usage{
make_aggregation_tbl(.data, date, value, time_unit)
}
\arguments{
\item{.data}{tibble}

\item{time_unit}{which time unit to aggregate to}

\item{...}{dimensions to group by}

\item{date_var}{dimension with main date column}

\item{value_var}{dimension with values to aggregate by}
}
\value{
tibble
}
\description{
Aggregate and expand date table
}
\examples{
make_aggregation_tbl(fpaR::sales,date=date,value=quantity,time_unit="day")
}
