% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_date.R
\name{make_cohort_tbl}
\alias{make_cohort_tbl}
\title{Make Cohort Table}
\usage{
make_cohort_tbl(.data, id_var, date, time_unit = "month", period_label = FALSE)
}
\arguments{
\item{time_unit}{}
}
\description{
Make Cohort Table
}
