% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{make_print_message}
\alias{make_print_message}
\title{Print message}
\usage{
make_print_message(x)
}
\arguments{
\item{x}{ti class object}
}
\value{
print message
}
\description{
Print message
}
\examples{
 x <- fpaR::ytd(fpaR::sales,date = order_date,value=quantity,calendar_type = "standard")
 make_print_message(x)
}
