% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_segmentation.R
\name{make_segmentation}
\alias{make_segmentation}
\title{Title}
\usage{
make_segmentation(
  .data,
  id_col,
  kmeans_centers_init,
  kmeans_nstart = 100,
  kmeans_iter.max = 100,
  centers_grid_range = 1:15,
  ...
)
}
\arguments{
\item{kmeans_centers_init}{}
}
\description{
Title
}
\examples{
make_segmentation(
.data =mtcars |>
tibble::rownames_to_column(var="column1"),
id_col=column1,
kmeans_nstart = 10,
kmeans_centers_init = 5,
kmeans_iter.max = 100,
centers_grid_range = 1:15)
}
