% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence_fn.R
\name{mom_fn}
\alias{mom_fn}
\title{month-over-month execution function}
\usage{
mom_fn(x)
}
\arguments{
\item{x}{ti object}
}
\value{
dbi object
}
\description{
\code{\link[=mom_fn]{mom_fn()}} is the function that is called by \code{\link[=mom]{mom()}} when passed through to \link{calculate}
}
\details{
This is internal non exported function that is nested in ti class and is called upon when the underlying function is called
by \link{calculate}
This will return a dbi object that can converted to a tibble object with \code{\link[dplyr:collect]{dplyr::collect()}}
}
\seealso{
\code{\link[=mom]{mom()}} for the function's intent
}
\keyword{internal}
