% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence_fn.R
\name{momtd_fn}
\alias{momtd_fn}
\title{Current year to date over previous year-to-date for tibble objects}
\usage{
momtd_fn(x)
}
\arguments{
\item{x}{ti object}
}
\value{
dbi object
}
\description{
\code{\link[=momtd_fn]{momtd_fn()}} is the function that is called by \code{\link[=momtd]{momtd()}} when passed through to \link{calculate}
}
\details{
This is internal non exported function that is nested in ti class and is called upon when the underlying function is called
by \link{calculate}
This will return a dbi object that can converted to a tibble object with \code{\link[dplyr:collect]{dplyr::collect()}}
}
\seealso{
\code{\link[=momtd]{momtd()}} for the function's intent
}
\keyword{internal}
