% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{momtd_tbl}
\alias{momtd_tbl}
\title{Current year to date over previous year-to-date for tibble objects}
\usage{
momtd_tbl(x)
}
\arguments{
\item{x}{ti_tbl}
}
\value{
tibble
}
\description{
Current year to date over previous year-to-date for tibble objects
}
