% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence_fn.R
\name{mtd_fn}
\alias{mtd_fn}
\title{Month-to-date execution function}
\usage{
mtd_fn(x)
}
\arguments{
\item{x}{ti object}
}
\value{
dbi object
}
\description{
\code{\link[=mtd_fn]{mtd_fn()}} is the function that is called by \code{\link[=mtd]{mtd()}} when passed through to \link{calculate}
}
\details{
This is internal non exported function that is nested in ti class and is called upon when the underlying function is called
by \link{calculate}
This will return a dbi object that can converted to a tibble object with \code{\link[dplyr:collect]{dplyr::collect()}}
}
\seealso{
\code{\link[=mtd]{mtd()}} for the function's intent
}
\keyword{internal}
