% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_vs_returning.R
\name{new_vs_returning}
\alias{new_vs_returning}
\title{new vs. returning variable analysis}
\usage{
new_vs_returning(df, group, date, new_label, returning_label, date_unit)
}
\arguments{
\item{df}{dataframe}

\item{group}{grouping logic}

\item{date}{date to sort by}

\item{new_label}{label of new variable}

\item{returning_label}{label of returning variable}

\item{date_unit}{data dimension}
}
\description{
new vs. returning variable analysis
}
