% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{pqtd_tbl}
\alias{pqtd_tbl}
\title{Previous quarter-to-date for tibble objects}
\usage{
pqtd_tbl(x)
}
\arguments{
\item{x}{ti_tbl}
}
\value{
tibble
}
\description{
Previous quarter-to-date for tibble objects
}
