% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{print_actions_steps}
\alias{print_actions_steps}
\title{Print action steps}
\usage{
print_actions_steps(x)
}
\arguments{
\item{x}{an S7 class}
}
\value{
cli messages
}
\description{
Print action steps
}
