% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{pwtd_tbl}
\alias{pwtd_tbl}
\title{Previous month-to-date for tibble objects}
\usage{
pwtd_tbl(x)
}
\arguments{
\item{x}{ti_tbl}
}
\value{
tibble
}
\description{
Previous month-to-date for tibble objects
}
