% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{pytd_tbl}
\alias{pytd_tbl}
\title{Previous year-to-date for tibble objects}
\usage{
pytd_tbl(x)
}
\arguments{
\item{x}{ti_tbl}
}
\value{
tibble
}
\description{
Previous year-to-date for tibble objects
}
