% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence_fn.R
\name{qoqtd_fn}
\alias{qoqtd_fn}
\title{Current quarter to date over previous quarter-to-date for tibble objects}
\usage{
qoqtd_fn(x)
}
\arguments{
\item{x}{ti object}
}
\value{
dbi object
}
\description{
\code{\link[=qoqtd_fn]{qoqtd_fn()}} is the function that is called by \code{\link[=qoqtd]{qoqtd()}} when passed through to \link{calculate}
}
\details{
This is internal non exported function that is nested in ti class and is called upon when the underlying function is called
by \link{calculate}
This will return a dbi object that can converted to a tibble object with \code{\link[dplyr:collect]{dplyr::collect()}}
}
\seealso{
\code{\link[=qoqtd]{qoqtd()}} for the function's intent
}
\keyword{internal}
