% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence_fn.R
\name{qtdopq_fn}
\alias{qtdopq_fn}
\title{quarter-to-date over previous quarter execution function}
\usage{
qtdopq_fn(x)
}
\arguments{
\item{x}{ti object}
}
\value{
dbi object
}
\description{
\code{\link[=qtdopq_fn]{qtdopq_fn()}} is the function that is called by \code{\link[=qtdopq]{qtdopq()}} when passed through to \link{calculate}
}
\details{
This is internal non exported function that is nested in ti class and is called upon when the underlying function is called
by \link{calculate}
This will return a dbi object that can converted to a tibble object with \code{\link[dplyr:collect]{dplyr::collect()}}
}
\seealso{
\code{\link[=qtdopq]{qtdopq()}} for the function's intent
}
\keyword{internal}
