% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{show_in_excel}
\alias{show_in_excel}
\title{Converts tibble to csv file and opens up in excel}
\usage{
show_in_excel(.data)
}
\arguments{
\item{.data}{tibble of data}
}
\value{
a temp csv file
}
\description{
Converts tibble to csv file and opens up in excel
}
\examples{
mtcars |>  show_in_excel()
}
