% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sql.R
\name{sql_query_select}
\alias{sql_query_select}
\title{sql_query_select}
\usage{
sql_query_select(
  .data,
  previous_query,
  select,
  from,
  where = NULL,
  group_by = NULL,
  having = NULL,
  window = NULL,
  order_by = NULL,
  limit = NULL,
  distinct = FALSE
)
}
\arguments{
\item{.data}{DBI object}

\item{previous_query}{Logical indicator if the from value should reference the previous .data steps}

\item{select}{the select values}

\item{from}{the from caluse}

\item{where}{the where clause}

\item{group_by}{the group by clause}

\item{having}{the having clause}

\item{window}{the window operation}

\item{order_by}{the order by caluse}

\item{limit}{the limit}

\item{distinct}{the distinct values}
}
\value{
sql and con object
}
\description{
sql_query_select
}
\examples{
sql_query_select(db_sales,previous_query=TRUE,select=sql(select *))
}
