% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{totalatd_dbi}
\alias{totalatd_dbi}
\title{Total since inception}
\usage{
totalatd_dbi(.data, ..., date_var, value_var)
}
\arguments{
\item{.data}{tibble of values}

\item{...}{optional columns to group by}

\item{date_var}{column with date var to aggregate by}

\item{value_var}{column with value to aggregate}
}
\description{
Total since inception
}
