% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sql.R
\name{with}
\alias{with}
\title{Helps creates CTE queries}
\usage{
with(query, query_name, order = c("middle"))
}
\arguments{
\item{query}{sql query}

\item{query_name}{CTE name}

\item{order}{CTE order either first, middle, last or single}
}
\value{
sql text object
}
\description{
Helps creates CTE queries
}
\examples{
query <- dplyr::sql("select * from sales")
with(query=query,query_name=sales_cte,order='first')
}
