% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{wow_dbi}
\alias{wow_dbi}
\title{week over week comparison for DBI objects}
\usage{
wow_dbi(.data, ..., date_var, value_var, lag_n = 1)
}
\arguments{
\item{.data}{tibble of values}

\item{...}{optional columns to group by}

\item{date_var}{column with date var to aggregate by}

\item{value_var}{column with value to aggregate}

\item{lag_n}{the number of weeks to lag}
}
\value{
DBI object
}
\description{
week over week comparison for DBI objects
}
