% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{abc}
\alias{abc}
\title{ABC classification}
\usage{
abc(df, group, dim, a = 0.7, b = 0.26, c = 0.04, na.rm = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{group}{grouping variable to classify}

\item{dim}{dimension to classify by (should be additive and positive)}

\item{a}{initial segment threshold expressed as a percentage}

\item{b}{incremental segment threshold expressed as a percentage}

\item{c}{final segment threshold expressed as a percentage}

\item{na.rm}{to remove na values}
}
\value{

}
\description{
ABC classification
}
\examples{

customer_abc <- abc(contoso,customer_key,margin,a=.7,b=.2,c=.1)
}
