% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{abc}
\alias{abc}
\title{ABC classification function}
\usage{
abc(.data, category_values, .value)
}
\arguments{
\item{.data}{tibble or dbi object (either grouped or ungrouped)}

\item{category_values}{vector of break points between 0 and 1}

\item{.value}{optional: if left blank will use the number of rows per group to category, alternatively you can pass a column name to categorize}
}
\value{
segment object
}
\description{
\itemize{
\item For your group variable, \code{\link[=abc]{abc()}}  will categorize which groups make up what proportion of the totals according to the category_values that you have entered
\item The function returns a segment object which prints out the summary of steps and actions that it will to categorize your data
\item Use \link{calculate} to return the results
}
}
\details{
\itemize{
\item This function xxxxx
\item If you provide a grouped object with \code{\link[dplyr:group_by]{dplyr::group_by()}}, it will generate a complete calendar for each group
\item The function creates a \code{segment} object, which pre-processes the data and arguments for further downstream functions
}
}
\examples{
abc(sales,c(.1,.5,.7,.96,1),.value=margin)
}
