% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{abc_fn}
\alias{abc_fn}
\title{Classify a group by proportion of a variable (A,B,C,...)}
\usage{
abc_fn(x)
}
\arguments{
\item{.value}{the column to aggregate}

\item{category_values}{the break points you want for your categories, each should be less than or equal to 1}

\item{fn}{'sum' to aggregate the value or 'n' to do the count of the group}

\item{.data}{tibble or DBI object}
}
\value{
a dbi objection
}
\description{
This returns a table that will segment your data into A,B or C segments based on custom
thresholds
}
