% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_graph.R
\name{abc_graph}
\alias{abc_graph}
\title{Produces ABC graph based on abc() function}
\usage{
abc_graph(.data)
}
\arguments{
\item{.data}{a tibble or DBI object producteed by abc() function}
}
\value{
ggplot
}
\description{
Produces ABC graph based on abc() function
}
\examples{
abc(ggplot2::diamonds,cut,dim=price,func="n") |> abc_graph()
}
