% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{calculate}
\alias{calculate}
\title{Total year to date aggregtation}
\usage{
calculate(.data, ..., date_var, value_var, time_unit)
}
\arguments{
\item{.data}{table}

\item{...}{dimensions to aggregrate by}

\item{date_var}{column with date var to aggregate by}

\item{value_var}{column with the value to aggregate}

\item{time_unit}{}
}
\value{
tibble
}
\description{
Total year to date aggregtation
}
