% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{Calculate to filter, summarize and pivot data}
\usage{
calculate(.data, .fn, rows, cols, filter)
}
\arguments{
\item{.data}{dataframe or tibble}

\item{.fn}{quote aggregation function}

\item{rows}{category in rows}

\item{cols}{category in cols}

\item{filter}{filter category, use & instead of ,}
}
\description{
Calculate to filter, summarize and pivot data
}
\examples{
calculate(ggplot2::diamonds,.fn="min(price)",rows=cut,cols=color,filter='color=="D"&cut=="Fair"')
}
