% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{calculate}
\alias{calculate}
\title{Calculate}
\arguments{
\item{x}{ti object}
}
\value{
dbi object
}
\description{
Calculate
}
\examples{
x <- ytd(sales,.date=order_date,.value=quantity,calendar_type="standard")
calculate(x)
}
