% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_plus.R
\name{count_plus}
\alias{count_plus}
\title{Proportion and Cumsum Portion}
\usage{
count_plus(.data, ..., .wt = FALSE, .sort = 1)
}
\arguments{
\item{.data}{dataframe}

\item{...}{columns you want to groupby}

\item{.wt}{if you want to weight it by a variable}

\item{.sort}{if you want to sort from largest (1), reverse(-1) or nothing 0}
}
\value{
a tibble of values
}
\description{
Proportion and Cumsum Portion
}
\examples{
ggplot2::diamonds \%>\% count_plus(color,cut,.wt=price)
}
