% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_plus.R
\name{count_plus}
\alias{count_plus}
\title{Dplyr's count() function but with cumulative totals}
\usage{
count_plus(df, dim, wt = NULL, sort = TRUE, round = 3)
}
\arguments{
\item{df}{dataframe}

\item{dim}{dimension to count}

\item{wt}{weighting dimension}

\item{sort}{sort results}

\item{round}{digits to round by}
}
\description{
Dplyr's count() function but with cumulative totals
}
