% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_date_tbl_sql.R
\name{create_date_sql}
\alias{create_date_sql}
\title{Create table of dates in sql}
\usage{
create_date_sql(.data, start_date, end_date)
}
\arguments{
\item{.data}{dbi object}

\item{start_date}{start date of date table range}

\item{end_date}{end date of date table range}
}
\value{
a sql query
}
\description{
Create a sql script of date table range for use in downstream queries
}
\examples{
create_date_sql(mtcars_dbi,start_date="2021-01-01",end_date="2022-01-01")
}
