% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.R
\name{divide}
\alias{divide}
\title{Divide function with error handling for divide by zero or NA}
\usage{
divide(numerator, denominator, alternative_result = NA_integer_)
}
\arguments{
\item{numerator}{numeric numerator}

\item{denominator}{numeric denominator}

\item{alternative_result}{alternative results if divide by 0 can by numeric or character but beware of class coercion}
}
\description{
Divide function with error handling for divide by zero or NA
}
\examples{
divide(100,0,"error")
}
