% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{dod}
\alias{dod}
\title{day over day calculation}
\usage{
dod(.data, date, value, calendar_type, lag_n = 1)
}
\arguments{
\item{.data}{tibble or DBI object}

\item{date}{the date column to aggregate}

\item{value}{the value column to summarize}

\item{calendar_type}{either 'standard' or '5-5-4' calendar}

\item{lag_n}{the number of periods to lag, default is 1}
}
\value{
dod_tbl or dbi_tbl
}
\description{
day over day calculation
}
\examples{
dod(fpaR::sales,date=order_date,value=quantity,calendar_type='standard',lag_n=1)
}
