% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{make_aggregation_dbi}
\alias{make_aggregation_dbi}
\title{Aggregate and expand date table}
\usage{
make_aggregation_dbi(.data, date, value, time_unit)
}
\arguments{
\item{.data}{dbi object}

\item{value}{}

\item{time_unit}{which time unit to aggregate to}

\item{...}{dimensions to group by}

\item{date_var}{dimension with main date column}

\item{value_var}{dimension with values to aggregate by}
}
\value{
list of dbi and sql object
}
\description{
Aggregate and expand date table
}
