% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{make_aggregation_tbl}
\alias{make_aggregation_tbl}
\title{Aggregate and expand date table}
\usage{
make_aggregation_tbl(.data, ..., date_var, value_var, time_unit)
}
\arguments{
\item{.data}{tibble}

\item{...}{dimensions to group by}

\item{date_var}{dimension with main date column}

\item{value_var}{dimension with values to aggregate by}

\item{time_unit}{which time unit to aggregate to}
}
\value{
tibble
}
\description{
Aggregate and expand date table
}
\examples{
make_aggregation_tbl(fpaR::sales,date,quantity,"day")
}
