% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{make_aggregation_tbl}
\alias{make_aggregation_tbl}
\title{Aggregate and expand date table}
\usage{
make_aggregation_tbl(.data, ..., date_var, value_var, time_unit)
}
\arguments{
\item{time_unit}{}
}
\value{
tibble
}
\description{
Aggregate and expand date table
}
\examples{
make_aggregation_tbl(fpaR::sales,date,quantity,"day")
}
