% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{make_aggregation_tbl}
\alias{make_aggregation_tbl}
\title{Aggregate and expand date table}
\usage{
make_aggregation_tbl(.data, date, value, time_unit)
}
\arguments{
\item{.data}{data either grouped or ungrouped}

\item{date}{the date column}

\item{value}{the value column to aggregate}

\item{time_unit}{the time unit to aggregate the date column by: 'day', 'week', 'month', 'quarter' or 'year'}
}
\value{
tibble
}
\description{
\code{make_aggregation_tbl()} summarizes a tibble to target time unit and completes the calendar to ensure
no missing days, month, quarter or years. If a grouped tibble is passed through it will complete the calendar
for each combination of the group
}
\details{
This is in internal function to make it easier to ensure data has no missing dates to
simplify the use of time intelligence functions downstream of the application.
If you want to summarize to a particular group, simply pass the tibble through to the \code{group_by()} argument
prior to function and the function will make summarize and make a complete calendar for each group item.
}
\examples{
make_aggregation_tbl(fpaR::sales,date=date,value=quantity,time_unit="day")
}
\seealso{
\code{\link[fpar:make_aggregation_dbi]{fpar::make_aggregation_dbi()}} which is the DBI equivalent of this function
}
