% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sql.R
\name{make_db_tbl}
\alias{make_db_tbl}
\title{Make an in memory database from a table}
\usage{
make_db_tbl(x)
}
\arguments{
\item{x}{tibble or dbi object}
}
\value{
dbi object
}
\description{
Make an in memory database from a table
}
