% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{make_print_message}
\alias{make_print_message}
\title{Make print message}
\usage{
make_print_message(x)
}
\arguments{
\item{x}{ti_tbl class object}
}
\value{
print message
}
\description{
Make print message
}
