% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_segmentation.R
\name{make_segmentation}
\alias{make_segmentation}
\title{Make a segmentation with kmeans and umap}
\usage{
make_segmentation(
  .data,
  id_col,
  kmeans_centers_init,
  kmeans_nstart = 100,
  kmeans_iter.max = 100,
  centers_grid_range = 1:15,
  max.overlaps = 10,
  ...
)
}
\arguments{
\item{.data}{a tibble dataframe with an ID column and numeric variables}

\item{id_col}{the ID column to join the kmeans and umap results together}

\item{kmeans_centers_init}{the k means center to start with}

\item{kmeans_nstart}{random variable to start iteration}

\item{kmeans_iter.max}{the maximum number of iterations}

\item{centers_grid_range}{range of kmeans center}

\item{...}{}
}
\value{
a list of objects including tibbles and ggplot
}
\description{
Make a segmentation with kmeans and umap
}
\examples{
make_segmentation(
.data =mtcars |>
tibble::rownames_to_column(var="column1"),
id_col=column1,
kmeans_nstart = 10,
kmeans_centers_init = 5,
kmeans_iter.max = 100,
centers_grid_range = 1:15)
}
