% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{mom}
\alias{mom}
\title{Month-over-month}
\usage{
mom(.data, .date, .value, calendar_type, lag_n = 1)
}
\arguments{
\item{.data}{tibble or DBI object}

\item{.date}{the .date column to aggregate}

\item{calendar_type}{either 'standard' or '5-5-4' calendar}

\item{lag_n}{the number of months to lag, default is 1}

\item{value}{the value column to summarize}
}
\value{
mom_tbl or mom_dbi
}
\description{
Month-over-month
}
\examples{
mom(fpaR::sales,.date=order_.date,.value=quantity,calendar_type='standard',lag_n=1)
}
