% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{mom_dbi}
\alias{mom_dbi}
\title{Month over month values}
\usage{
mom_dbi(.data, ..., date_var, value_var, lag_n = 1)
}
\arguments{
\item{.data}{tibble of values}

\item{...}{optional columns to group by}

\item{date_var}{column with date var to aggregate by}

\item{value_var}{column with value to aggregate}

\item{lag_n}{the number of months to lag}
}
\description{
Month over month values
}
