% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{momtd}
\alias{momtd}
\title{Month-to-.date}
\usage{
momtd(.data, .date, .value, calendar_type, lag_n)
}
\arguments{
\item{.data}{either a tibble or  DBI object}

\item{.date}{the .date column to aggregate}

\item{calendar_type}{either 'standard' or '5-5-4' calendar}

\item{value}{the value column to summarize}
}
\value{
mtd_tbl or mtd_dbi object depending on what is passed through
}
\description{
This calculates the monthly cumulative sum of targeted value using a standard or 5-5-4 calendar respecting
any groups that are passed through with \code{dplyr::group_by()}.

Use \code{calculate()} to return the results
}
\examples{
mtd(fpaR::sales,.date=.date,.value=quantity,calendar_type="standard")
}
