% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{pmtd_fn}
\alias{pmtd_fn}
\title{Previous month-to-date for tibble objects}
\usage{
pmtd_fn(x)
}
\arguments{
\item{x}{ti_tbl}
}
\value{
tibble
}
\description{
Previous month-to-date for tibble objects
}
