% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence_fn.R
\name{pmtd_fn}
\alias{pmtd_fn}
\title{Previous month-to-date for tibble objects}
\usage{
pmtd_fn(x)
}
\arguments{
\item{x}{ti object}
}
\value{
dbi object
}
\description{
\code{pmtd_fn()} is the function that is called by \code{pmtd()} when passed through to \code{calculate()}
}
\details{
This is internal non exported function that is nested in ti class and is called upon when the underlying function is called
by \code{calculated()}
This will return a dbi object that can converted to a tibble object with \code{dplyr::collect()}
}
\seealso{
\code{\link[=pmtd]{pmtd()}} for the function's intent
}
