% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{pqtd}
\alias{pqtd}
\title{Previous quarter quarter-to-date}
\usage{
pqtd(.data, date, value, calendar_type, lag_n)
}
\arguments{
\item{.data}{either a tibble or  DBI object}

\item{date}{the date column to aggregate}

\item{value}{the value column to summarize}

\item{calendar_type}{either 'standard' or '5-5-4' calendar}
}
\value{
ytd_tbl or ytd_dbi
}
\description{
This calculates the annual cumulative sum of targeted value using a standard or 5-5-4 calendar respecting
any groups that are passed through with \code{dplyr::group_by()}.

Use \code{calculate()} to return the results
}
\examples{
ytd(fpaR::sales,date=date,value=quantity,calendar_type="standard")
}
