% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence_fn.R
\name{pqtd_fn}
\alias{pqtd_fn}
\title{Previous quarter-to-date for tibble objects}
\usage{
pqtd_fn(x)
}
\arguments{
\item{x}{ti object}
}
\value{
dbi object
}
\description{
\code{\link[=pqtd_fn]{pqtd_fn()}} is the function that is called by \code{\link[=pqtd]{pqtd()}} when passed through to \link{calculate}
}
\details{
This is internal non exported function that is nested in ti class and is called upon when the underlying function is called
by \link{calculate}
This will return a dbi object that can converted to a tibble object with \code{\link[dplyr:compute]{dplyr::collect()}}
}
\seealso{
\code{\link[=pqtd]{pqtd()}} for the function's intent
}
