% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print}
\alias{print}
\title{Print ti objects}
\arguments{
\item{x}{ti object}

\item{...}{additional arguments}
}
\value{
ti object
}
\description{
Print ti objects
}
\examples{
x <- ytd(sales,.date=order_date,.value=quantity,calendar_type="standard")
x
}
