% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print}
\alias{print}
\alias{print.ti}
\title{Print ti objects}
\usage{
\method{print}{ti}(x, ...)
}
\arguments{
\item{x}{ti object}

\item{...}{unused. Please ignore.}
}
\value{
ti object
}
\description{
Print ti objects
}
