% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{pwtd}
\alias{pwtd}
\title{Week-to-date}
\usage{
pwtd(.data, date, value, calendar_type, lag_n)
}
\arguments{
\item{.data}{either a tibble or  DBI object}

\item{date}{the date column to aggregate}

\item{value}{the value column to summarize}

\item{calendar_type}{either 'standard' or '5-5-4' calendar}
}
\value{
wtd_tbl or wtd_dbi
}
\description{
This calculates the weekly cumulative sum of targeted value using a standard or 5-5-4 calendar respecting
any groups that are passed through with \code{dplyr::group_by()}.

Use \code{calculate()} to return the results
}
\examples{
wtd(fpaR::sales,date=date,value=quantity,calendar_type="standard")
}
