% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{pytd_fn}
\alias{pytd_fn}
\title{Previous year-to-date for tibble objects}
\usage{
pytd_fn(x)
}
\arguments{
\item{x}{ti object}
}
\value{
dbi

tibble
}
\description{
\code{pytd_fn()} summarizes a tibble to target time unit and completes the calendar to ensure
no missing days, month, quarter or years. If a grouped tibble is passed through it will complete the calendar
for each combination of the group
}
\details{
This is in internal function to make it easier to ensure data has no missing dates to
simplify the use of time intelligence functions downstream of the application.
If you want to summarize to a particular group, simply pass the tibble through to the \code{group_by()} argument
prior to function and the function will make summarize and make a complete calendar for each group item.
}
