% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence_classes.R
\name{qoq}
\alias{qoq}
\title{Current full period quarter over previous full period quarter}
\usage{
qoq(.data, .date, .value, calendar_type, lag_n = 1)
}
\arguments{
\item{.data}{tibble or dbi object (either grouped or ungrouped)}

\item{.date}{the date column to group by}

\item{.value}{the value column to summarize}

\item{calendar_type}{select either 'standard' or '5-5-4' calendar, see 'Details' for additional information}

\item{lag_n}{the number of periods to lag}
}
\value{
ti object
}
\description{
A short description...
}
\examples{
qoq(sales,.date=order_date,.value=quantity,calendar_type='standard',lag_n=1)
}
\seealso{
Other time_intelligence: 
\code{\link{atd}()},
\code{\link{dod}()},
\code{\link{mom}()},
\code{\link{momtd}()},
\code{\link{mtd}()},
\code{\link{mtdopm}()},
\code{\link{pmtd}()},
\code{\link{pqtd}()},
\code{\link{pwtd}()},
\code{\link{pytd}()},
\code{\link{qoqtd}()},
\code{\link{qtd}()},
\code{\link{qtdopq}()},
\code{\link{wow}()},
\code{\link{wowtd}()},
\code{\link{wtd}()},
\code{\link{wtdopw}()},
\code{\link{yoy}()},
\code{\link{yoytd}()},
\code{\link{ytd}()},
\code{\link{ytdopy}()}
}
\concept{time_intelligence}
