% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{qoq_fn}
\alias{qoq_fn}
\title{Quarter-over-quarter execution function}
\usage{
qoq_fn(x)
}
\arguments{
\item{x}{ti object}
}
\value{
dbi object
}
\description{
Quarter-over-quarter execution function
}
