% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{qoqtd_fn}
\alias{qoqtd_fn}
\title{Current quarter to date over previous quarter-to-date for tibble objects}
\usage{
qoqtd_fn(x)
}
\arguments{
\item{x}{ti object}
}
\value{
dbi object
}
\description{
\code{qoqtd_fn()} is the function that is called by \code{qoqtd()} when passed through to \code{calculate()}
}
\details{
This is internal non exported function that is nested in ti class and is called upon when the underlying function is called
by \code{calculated()}
This will return a dbi object that can converted to a tibble object with \code{dplyr::collect()}
}
\seealso{
\code{\link[=qoqtd]{qoqtd()}} for the function's intent
}
