% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{qoqtd_fn}
\alias{qoqtd_fn}
\title{Previous quarter-to-date for tibble objects}
\usage{
qoqtd_fn(x)
}
\arguments{
\item{x}{ti_tbl}
}
\value{
tibble
}
\description{
Previous quarter-to-date for tibble objects
}
