% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{qtd}
\alias{qtd}
\title{quarter-to-.date}
\usage{
qtd(.data, .date, .value, calendar_type)
}
\arguments{
\item{.data}{either a tibble or  dbi object}

\item{.date}{the .date column to aggregate}

\item{calendar_type}{either 'standard' or '5-5-4' calendar}

\item{value}{the value column to summarize}
}
\value{
qtd_tbl or qtd_dbi object depending on what is passed through
}
\description{
this calculates the quarterly cumulative sum of targeted value using a standard or 5-5-4 calendar respecting
any groups that are passed through with \code{dplyr::group_by()}.

use \code{calculate()} to return the results
}
\examples{
qtd(fpar::sales,.date=.date,.value=quantity,calendar_type="standard")
}
