% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{qtdopq_fn}
\alias{qtdopq_fn}
\title{quarter-to-date over previous quarter}
\usage{
qtdopq_fn(x)
}
\arguments{
\item{x}{ti_tbl}
}
\value{
tibble
}
\description{
quarter-to-date over previous quarter
}
