% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_in_excel.R
\name{show_in_excel}
\alias{show_in_excel}
\title{Converts dataframe to csv file and opens up in excel}
\usage{
show_in_excel(.data)
}
\arguments{
\item{.data}{dataframe in tibble or dataframe format}
}
\value{
a csv file
}
\description{
Converts dataframe to csv file and opens up in excel
}
\examples{
mtcars |>  show_in_excel()
}
