% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{totalmtd}
\alias{totalmtd}
\title{Total month to date values}
\usage{
totalmtd(.data, ..., date_var, value_var)
}
\arguments{
\item{.data}{tibble of values}

\item{...}{optional columns to group by}

\item{date_var}{column with date var to aggregate by}

\item{value_var}{column with value to aggregate}
}
\value{
tibble
}
\description{
Total month to date values
}
\examples{
totalmtd(sales_tbl,date_var = order_date,value_var = quantity)
}
