% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{totalwtd_dbi}
\alias{totalwtd_dbi}
\title{Total week to date values}
\usage{
totalwtd_dbi(.data, ..., date_var, value_var)
}
\arguments{
\item{.data}{tibble of values}

\item{...}{optional columns to group by}

\item{date_var}{column with date var to aggregate by}

\item{value_var}{column with value to aggregate}
}
\description{
Total week to date values
}
