% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{wowtd_fn}
\alias{wowtd_fn}
\title{Current year to date over previous year-to-date for tibble objects}
\usage{
wowtd_fn(x)
}
\arguments{
\item{x}{ti_tbl}
}
\value{
tibble
}
\description{
Current year to date over previous year-to-date for tibble objects
}
