% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{ytd}
\alias{ytd}
\title{Year-to-date}
\usage{
ytd(.data, date, value, calendar_type)
}
\arguments{
\item{.data}{either a dataframe or lazy DBI object}

\item{date}{the date column to aggregate}

\item{value}{the value column to summarize}

\item{calendar_type}{either 'standard' or '5-5-4' calendar}
}
\description{
Year-to-date
}
