% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{ytd}
\alias{ytd}
\title{Year-to-date}
\usage{
ytd(.data, .date, .value, calendar_type)
}
\arguments{
\item{.data}{either a tibble or  DBI object}

\item{calendar_type}{either 'standard' or '5-5-4' calendar}

\item{date}{the date column to aggregate}

\item{value}{the value column to summarize}
}
\value{
ytd_tbl or ytd_dbi
}
\description{
This calculates the annual cumulative sum of targeted value using a standard or 5-5-4 calendar respecting
any groups that are passed through with \code{dplyr::group_by()}.

Use \code{calculate()} to return the results
}
\examples{
ytd(fpaR::sales,date=date,value=quantity,calendar_type="standard")
}
