% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{ytd}
\alias{ytd}
\title{Year-to-date}
\usage{
ytd(.data, .date, .value, calendar_type)
}
\arguments{
\item{.data}{tibble or dbi object}

\item{.date}{the date column to group by}

\item{.value}{the value column to summarize}

\item{calendar_type}{either 'standard' or '5-5-4' calendar}
}
\value{
ti object
}
\description{
For each group, \code{ytd()}  will the running sum of a value based on the calendar date.
This function will print out the summary of steps and actions. Use \code{calculate()} to return the results
}
\details{
This creates a calendar object with complete dates to ensure no missing days, month, quarter or years.
If a grouped object is passed through it will complete the calendar  for each combination of the group.
This is part of the time intelligence family of functions.
Pass the arguments to  \code{ytd()} and it will print out its actions and assumptions
Use \code{calculate()} to return the results
}
\examples{
ytd(fpaR::sales,.date=order_date,.value=quantity,calendar_type="standard")
}
\seealso{
\code{\link[=pytd]{pytd()}}, \code{\link[=yoy]{yoy()}}, \code{\link[=yoytd]{yoytd()}} are the year-based ti functions
}
