% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_intelligence.R
\name{ytdopy_fn}
\alias{ytdopy_fn}
\title{Year-to-date vs. full prior year}
\usage{
ytdopy_fn(x)

ytdopy_fn(x)
}
\arguments{
\item{x}{ti_tbl}
}
\value{
tibble

tibble
}
\description{
Year-to-date vs. full prior year

quarter-to-date over previous quarter
}
